Implementacion de autorizacion usando Gates

+++ app/Http/Controllers/QuestionController.php
@@ -71,6 +71,10 @@ class QuestionController extends Controller
     public function edit(Question $question)
     {
         //Laravel inyecta el modelo completo
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Acceso denegado");    
+        }
+        
         return view("questions.edit", compact('question'));
     }
 
@@ -83,6 +87,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Acceso denegado");    
+        }
         $question->update($request->only('title', 'body'));
 
         return \redirect('/questions')->with('success', 'Su pregunta se ha actualizado');
@@ -96,6 +103,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Acceso denegado");    
+        }
         $question->delete();
         
         return redirect('/questions')->with('success', 'Pregunta ha sido eliminada');
